Implement "original file" mode
This gives access to the original PDF and EPUB files.  Notes are
given access to the raw zip files at this point.  To get the file
type and size, we request the end of blob, which has the zip file
central directory.  This lets us read through all of data about
the files in the zip.  It takes some guessing, which isn't too
clever yet.  First, we guess that the directory is contained in the
last 100 KB of the file.  A rough estimate suggests that this will
be okay for documents of up to 500 pages.  Second, we guess that
the only place we'll see the string '.content' is in the first
header in the central directory, so we back up a fixed distance from
there and hope that's the beginning.