Add a simple cache of document contents
Previously, we held the contents in memory forever, which is a bad
idea.  But we do want some caching, since we will often get a number
of short reads of a single file.  Mainly, this commit introduces an
API to use for future experiments.  But here we have a very simple
cache that holds only the last file in memory.