Switch to async HTTP operations
Since pyfuse3 supports it, we might as well keep from blocking on the
many HTTP requests we're sending.  asks is mostly a drop in replacement
for requests, so that was straightforward.

We had exposed things like the raw file and content as properties, which
hid the fact that several HTTP requests might be happening to get their
values.  While it does seem like async properties can be a thing, since
we were touching all of these things anyways, I went ahead and changed
them to async methods.

Since multiple requests can be in flight at once, it could happen that
something would trigger a request that's already in flight.  I've added
some locking to try to prevent this.  Note that it's not enough to just
lock the immediate request -- we have to lock enough around that to
ensure that we don't decide that a request is necessary when one is
already outstanding.  In the client, this is pretty obvious.  We just
lock the large update operation.  In the items, there are several
requests that could go out: getting the download URL, getting the
raw file size, getting the file details from the end of the download,
and getting the full raw download.  It wasn't immediately clear to
me how all of these should interact, so right now a single lock
guards all of them.  This might be optimizable later.  Also, some of
these things can end up needing others inside of them.  The with_lock
decorator checks if we already hold this lock before requesting it.