Add FS support for creating files
This logic gets spread out over a few functions.  create is called,
which should both create an inode and a file handle.  (Note there's
also mknod, which just creates the inode.  In simple testing, create
is tried first, then mknod.  Since we want to only create a file if
we immediate write to it, I'm hoping this is enough, and mknod isn't
implemented yet.)  Then the file handle is written to multiple times.
The data is stored in a buffer.  When the file is closed, release gets
called.  It's at this point that we have the full file and can upload
it to the cloud.

Unfortunately, errors in release are ignored (since the writes have
already succeeded).  Thus, if something goes wrong in uploading the
file, nothing gets conveyed back the the user.  There is a flush
call that comes between the last write and the release.  The
documentation notes that this can be called multiple times, if there
were multiple copies of the file handle, whereas release will only
be called once.  It's unclear if flush does anything more useful
with errors.