Handle items with missing parents
It's possible to get an item with a parent ID that doesn't exist.
(See rmfuse #24, for instance.)  Remarkable just puts such an item
in the root directory, so we'll do that too.  This code also checks
that the parent is a folder, not a document, and treats that case
similarly.  (I don't know what Remarkable does in that case.)